requirements.txt
Flask==2.3.2
opencv_python==4.7.0.72
Pillow==9.5.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
